What is the Monty Hall Problem?
The Monty Hall problem is a counter-intuitive statistics puzzle:

There are 3 doors, behind which are two goats and a car.
You pick a door (call it door A). You’re hoping for the car of course.
Monty Hall, the game show host, examines the other doors (B & C) and opens one with a goat. (If both doors have goats, he picks randomly.)
Here’s the game: Do you stick with door A (original guess) or switch to the unopened door? Does it matter?

Surprisingly, the odds aren’t 50-50. If you switch doors you’ll win 2/3 of the time! [Source](https://betterexplained.com/articles/understanding-the-monty-hall-problem/)

1. Download the code
2. Open the file using your local python editor
3. Run the simulation